chore: move references to separate file
